+++ classifier/ClassifierObject.h
@@ -3,9 +3,9 @@
 class ClassifierObject
 {
 public:
-	float *data;
+	double *data;
 	virtual ~ClassifierObject();
 	virtual int getColumnCount() = 0;
-	void setDataAt(int i, float d);
+	void setDataAt(int i, double d);
 };
